% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSum.R
\name{FlowSum}
\alias{FlowSum}
\title{Flow Sum Calculation}
\usage{
FlowSum(df, df.names = c("SITEYR", "ChannelNum", "Lat_Loc_m", "Depth_m",
  "Velocity_m_s"), returnType = "sample")
}
\arguments{
\item{df}{Data frame of data to be processed.  Must include the columns specified in the parameter df1.names.}

\item{df.names}{Column names; Optional. Defaults = c("SITEYR", "ChannelNum", "Lat_Loc_m", "Depth_m", "Velocity_m_s")}

\item{returnType}{Returns the original data with cell velocity (cell) or a summary table of the sample velocities (sample); "sample" is the default.}
}
\value{
Returns the requested dataframe; the original data with the area velocity by cell or by sample (default).
}
\description{
Calculate flow from field measurements.  All measurements are in meters and meters per second.
The final flow measurement is reports as cms and cubic feet per second (cfs).
}
\examples{
#open data
data.flow <-MBSS.flow
#data.flow <- load("./data/MBSS.flow.rda")
# names
# calculate flow
flow.cell <- FlowSum(data.flow,returnType="cell")
flow.sample <- FlowSum(data.flow)
# examine data
View(flow.cell)
View(flow.sample)
# Save file (tab delimited text file)
datetime <- format(Sys.time(),"\%Y\%m\%d_\%H\%M\%S")
myYear <- "15"
write.table(flow.sample,paste0("SumFlow",myYear,"_",datetime,".tab"),row.names=FALSE,sep="\\t")
}
\keyword{MBSS,}
\keyword{flow}

