% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{MBSS.flow}
\alias{MBSS.flow}
\title{MBSS.flow}
\format{
a data frame with 2,686 observations and 5 variables:
}
\source{
Example flow data for 2015
}
\usage{
MBSS.flow
}
\description{
MBSS flow data.
}
\keyword{datasets}
