% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{MBSS.Ion.Data}
\alias{MBSS.Ion.Data}
\title{MBSS.Ion.Data}
\format{
a data frame with 41 observations and 14 variables:
}
\source{
December 2106 data
}
\usage{
MBSS.Ion.Data
}
\description{
MBSS ion example data.
}
\keyword{datasets}
