% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{MBSS.Ion.Ref}
\alias{MBSS.Ion.Ref}
\title{MBSS.Ion.Ref}
\format{a data frame with 13 observations and 8 variables:}
\source{
Reference values for ion contribution calculations.
}
\usage{
MBSS.Ion.Ref
}
\description{
MBSS ion reference data.  Used for calculating percent contribution of an ion to overall sample conductivity.
}
\keyword{datasets}

