% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.scores.R
\name{metric.scores}
\alias{metric.scores}
\title{Score metrics}
\usage{
metric.scores(DF_Metrics, MetricNames, IndexName, IndexRegion, DF_Thresh)
}
\arguments{
\item{DF_Metrics}{Data frame of metric values (as columns), Index Name, and Index Region (strata).}

\item{MetricNames}{Names of columns of metric values.}

\item{IndexName}{Nam of column with index (e.g., MBSS.2005.Bugs)}

\item{IndexRegion}{Name of column with relevant bioregion or site class (e.g., COASTAL).}

\item{DF_Thresh}{Data frame of Scoring Thresholds (Index.Name, Index.Region, Metric, Direction, Thresh.Lo, Thresh.Hi, ScoreRegime)}
}
\value{
vector of scores
}
\description{
This function calculates metric scores based on a Thresholds data frame.
Can generate scores for 1/3/5 (ScoreRegime="135") or 0-100 (ScoreRegime="100").
}
\examples{
# Metrics, Fish
#(generate values then score)s
myIndex <- "MBSS.2005.Fish"
# Thresholds
thresh <- metrics_scoring
# get metric names for myIndex
(myMetrics.Fish <- as.character(droplevels(unique(thresh[thresh[,"Index.Name"]==myIndex,"Metric"]))))
# Taxa Data
myDF.Fish <- taxa_fish
myMetric.Values.Fish <- metric.values(myDF.Fish, "fish", myMetrics.Fish, TRUE)
View(myMetric.Values.Fish)
# SCORE
Metrics.Fish.Scores <- metric.scores(myMetric.Values.Fish, myMetrics.Fish, "Index.Name", "FIBISTRATA", thresh)
# View Results
View(Metrics.Fish.Scores)

# Metrics, Index, Benthic Macroinvertebrates, genus
# (generate values then scores)
myIndex <- "MBSS.2005.Bugs"
# Thresholds
thresh <- metrics_scoring
# get metric names for myIndex
(myMetrics.Bugs <- as.character(droplevels(unique(thresh[thresh[,"Index.Name"]==myIndex,"Metric"]))))
# Taxa Data
myDF.Bugs <- taxa_bugs_genus
myMetric.Values.Bugs <- metric.values(myDF.Bugs, "bugs", myMetrics.Bugs)
View(myMetric.Values.Bugs)
# SCORE
Metrics.Bugs.Scores <- metric.scores(myMetric.Values.Bugs, myMetrics.Bugs, "INDEX.NAME", "STRATA_R", thresh)
# View Results
View(Metrics.Bugs.Scores)

# QC bug count
Metrics.Bugs.Scores[Metrics.Bugs.Scores[,"totind"]>120,"QC_Count"] <- "LARGE"
Metrics.Bugs.Scores[Metrics.Bugs.Scores[,"totind"]<60,"QC_Count"] <- "SMALL"
Metrics.Bugs.Scores[is.na(Metrics.Bugs.Scores[,"QC_Count"]),"QC_Count"] <- "OK"
# table of QC_Count
table(Metrics.Bugs.Scores$QC_Count)
}
