% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.values.R
\name{metric.values.fish}
\alias{metric.values.fish}
\title{Calculate Metric Values for Fish}
\usage{
metric.values.fish(myDF, SampleID, MetricNames = NULL, boo.Adjust)
}
\arguments{
\item{myDF}{Data frame of taxa (list required fields)}

\item{SampleID}{Sample ID column}

\item{MetricNames}{Optional vector of metric names to be returned.  If
none are supplied then all will be returned.}

\item{boo.Adjust}{Optional boolean value on whether to perform adjustments of
values prior to scoring.  Default = FALSE but will always be TRUE for fish
metrics.}
}
\value{
data frame of SampleID and metric values to metric.values()
}
\description{
Subfunction of metric.values()
}
\examples{
# None

}
