% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{metrics_scoring}
\alias{metrics_scoring}
\title{metrics_scoring}
\format{a dataframe with 69 observations and 8 variables:}
\source{
Reference scoring thresholds for BIBI and FIBI, Southerland et al. 2005.
}
\usage{
metrics_scoring
}
\description{
MBSS IBI scoring information.  Used for scoring metrics in BIBI and FIBI.
}
\keyword{datasets}
