% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{run MBSStools Shiny Example}
\usage{
runShiny()
}
\description{
Launches Shiny app for MBSStools.
}
\details{
The Shiny app based on the R package MBSStools is included in the R
package. This function launches that app.

The Shiny app is online at:
hhttps://tetratech-wtr-wne.shinyapps.io/MBSStools/
}
\examples{
\dontrun{
# Run Function
runShiny()
}
}
