% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{shp_strata_bugs}
\alias{shp_strata_bugs}
\title{GIS shapefile, MBSS strata bugs}
\format{
Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots
\describe{
..@ data       :'data.frame':	3 obs. of  4 variables:
.. ..$ AREA      : num [1:3] 2.02e+10 0.00 0.00
.. ..$ PERIMETER : num [1:3] 708922 0 0
.. ..$ PROVINCES_: int [1:3] 6 1 2
.. ..$ Strata    : chr [1:3] "Coastal" "Piedmont" "Highlands"
..@ polygons   :List of 3
.. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
.. .. .. ..@ Polygons :List of 1
.. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
.. .. .. .. .. .. ..@ labpt  : num [1:2] -76.2 38.6
.. .. .. .. .. .. ..@ area   : num 2.09
.. .. .. .. .. .. ..@ hole   : logi FALSE
.. .. .. .. .. .. ..@ ringDir: int 1
.. .. .. .. .. .. ..@ coords : num [1:9585, 1:2] -77.1 -77.1 -77.1 -77.1 -77.1 ...
.. .. .. ..@ plotOrder: int 1
.. .. .. ..@ labpt    : num [1:2] -76.2 38.6
.. .. .. ..@ ID       : chr "0"
.. .. .. ..@ area     : num 2.09
.. .. .. ..$ comment: chr "0"
.. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
.. .. .. ..@ Polygons :List of 1
.. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
.. .. .. .. .. .. ..@ labpt  : num [1:2] -76.7 39.4
.. .. .. .. .. .. ..@ area   : num 0.498
.. .. .. .. .. .. ..@ hole   : logi FALSE
.. .. .. .. .. .. ..@ ringDir: int 1
.. .. .. .. .. .. ..@ coords : num [1:5258, 1:2] -75.8 -75.8 -75.8 -75.8 -75.8 ...
.. .. .. ..@ plotOrder: int 1
.. .. .. ..@ labpt    : num [1:2] -76.7 39.4
.. .. .. ..@ ID       : chr "1"
.. .. .. ..@ area     : num 0.498
.. .. .. ..$ comment: chr "0"
.. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
.. .. .. ..@ Polygons :List of 1
.. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
.. .. .. .. .. .. ..@ labpt  : num [1:2] -78.1 39.5
.. .. .. .. .. .. ..@ area   : num 0.738
.. .. .. .. .. .. ..@ hole   : logi FALSE
.. .. .. .. .. .. ..@ ringDir: int 1
.. .. .. .. .. .. ..@ coords : num [1:22165, 1:2] -76.9 -76.9 -76.9 -76.9 -76.9 ...
.. .. .. ..@ plotOrder: int 1
.. .. .. ..@ labpt    : num [1:2] -78.1 39.5
.. .. .. ..@ ID       : chr "2"
.. .. .. ..@ area     : num 0.738
.. .. .. ..$ comment: chr "0"
..@ plotOrder  : int [1:3] 1 3 2
..@ bbox       : num [1:2, 1:2] -79.5 37.9 -75 39.7
.. ..- attr(*, "dimnames")=List of 2
.. .. ..$ : chr [1:2] "x" "y"
#' .. .. ..$ : chr [1:2] "min" "max"
..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
.. .. ..@ projargs: chr "+proj=longlat +datum=WGS84 +no_defs"
.. .. ..$ comment: chr "GEOGCRS[\"unknown\",    DATUM[\"World Geodetic System 1984\",        ELLIPSOID[\"WGS 84\",6378137,298.25722"| __truncated__
..$ comment: chr "TRUE"
}
}
\source{
MBSS staff, mbss_strata.shp
}
\usage{
shp_strata_bugs
}
\description{
A shapefile with polygons for Maryland for MBSS strata for bugs.
Used with strata function.
}
\keyword{datasets}
