% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{shp_strata_msw}
\alias{shp_strata_msw}
\title{GIS shapefile, MBSS (MSW) strata, bugs}
\format{
{
The format is:
 Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots
..@ data       :'data.frame':	2 obs. of  4 variables:
 .. ..$ AREA      : num [1:2] 2.02e+10 1.18e+10
.. ..$ PERIMETER : num [1:2] 709260 989089
.. ..$ PROVINCES_: chr [1:2] "-2" "-1"
.. ..$ Strata    : chr [1:2] "CP" "NCP"
..@ polygons   :List of 2
.. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
.. .. .. ..@ Polygons :List of 1
.. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
.. .. .. .. .. .. ..@ labpt  : num [1:2] -76.2 38.6
.. .. .. .. .. .. ..@ area   : num 2.09
.. .. .. .. .. .. ..@ hole   : logi FALSE
.. .. .. .. .. .. ..@ ringDir: int 1
.. .. .. .. .. .. ..@ coords : num [1:9585, 1:2] -77.1 -77.1 -77.1 -77.1 -77.1 ...
.. .. .. ..@ plotOrder: int 1
.. .. .. ..@ labpt    : num [1:2] -76.2 38.6
.. .. .. ..@ ID       : chr "0"
.. .. .. ..@ area     : num 2.09
.. .. .. ..$ comment: chr "0"
.. ..$ :Formal class 'Polygons' [package "sp"] with 5 slots
.. .. .. ..@ Polygons :List of 1
.. .. .. .. ..$ :Formal class 'Polygon' [package "sp"] with 5 slots
.. .. .. .. .. .. ..@ labpt  : num [1:2] -77.5 39.5
.. .. .. .. .. .. ..@ area   : num 1.24
.. .. .. .. .. .. ..@ hole   : logi FALSE
.. .. .. .. .. .. ..@ ringDir: int 1
.. .. .. .. .. .. ..@ coords : num [1:24274, 1:2] -79 -79 -79 -79 -79 ...
.. .. .. ..@ plotOrder: int 1
.. .. .. ..@ labpt    : num [1:2] -77.5 39.5
.. .. .. ..@ ID       : chr "1"
.. .. .. ..@ area     : num 1.24
.. .. .. ..$ comment: chr "0"
..@ plotOrder  : int [1:2] 1 2
..@ bbox       : num [1:2, 1:2] -79.5 37.9 -75 39.7
.. ..- attr(*, "dimnames")=List of 2
.. .. ..$ : chr [1:2] "x" "y"
.. .. ..$ : chr [1:2] "min" "max"
..@ proj4string:Formal class 'CRS' [package "sp"] with 1 slot
.. .. ..@ projargs: chr "+proj=longlat +datum=WGS84 +no_defs"
.. .. ..$ comment: chr "GEOGCRS[\"unknown\",    DATUM[\"World Geodetic System 1984\",        ELLIPSOID[\"WGS 84\",6378137,298.25722"| __truncated__
..$ comment: chr "TRUE"
}
}
\source{
MBSS, manipulated mbss_strata for only two strata (NCP and CP).
}
\usage{
shp_strata_msw
}
\description{
A shapefile with polygons for Maryland for MBSS strata for MSW.
Used with strata function.
}
\keyword{datasets}
