% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{taxa_bugs_family}
\alias{taxa_bugs_family}
\title{taxa_bugs_family}
\format{a dataframe with 999 observations and 999 variables:}
\source{
MSW 2016 benthic macroinvertebrate family data.
}
\usage{
taxa_bugs_family
}
\description{
MSW benthic macroinvertebrate family taxa information from 2016.  Used as example for calculating metrics and index.
}
\keyword{datasets}
