% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{taxa_bugs_genus}
\alias{taxa_bugs_genus}
\title{taxa_bugs_genus}
\format{a dataframe with 5066 observations and 20 variables:}
\source{
MBSS 2016 benthic macroinvertebrate genus data.
}
\usage{
taxa_bugs_genus
}
\description{
MBSS benthic macroinvertebrate genus taxa information from 2016.  Used as example for calculating metrics and index.
}
\keyword{datasets}
