% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{taxa_fish}
\alias{taxa_fish}
\title{taxa_fish}
\format{
a dataframe with 1694 observations and 15 variables:
}
\source{
MBSS 2016 fish data.
}
\usage{
taxa_fish
}
\description{
MBSS fish taxa information from 2016.  Used as example for calculating
metrics and index.
One additional "No Fish Observed" sampled added for testing.
}
\keyword{datasets}
