% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{MBSS.PHI}
\alias{MBSS.PHI}
\title{MBSS.PHI}
\format{
a data frame with 81 observations and 20 variables:
}
\source{
PHI worksheet (Excel) master file from Paul et al. 2003_MBSS (ea-03-4_phi).
}
\usage{
MBSS.PHI
}
\description{
MBSS Physical Habitat Index (PHI) data.
}
\keyword{datasets}
