% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.scores.R
\name{metric.scores}
\alias{metric.scores}
\title{Score metrics}
\usage{
metric.scores(DF_Metrics, MetricNames, IndexName, IndexRegion, DF_Thresh)
}
\arguments{
\item{DF_Metrics}{Data frame of metric values (as columns), Index Name, and Index Region (strata).}

\item{MetricNames}{Names of columns of metric values.}

\item{IndexName}{Nam of column with index (e.g., MBSS.2005.Bugs)}

\item{IndexRegion}{Name of column with relevant bioregion or site class (e.g., COASTAL).}

\item{DF_Thresh}{Data frame of Scoring Thresholds (Index.Name, Index.Region, Metric, Direction, Thresh.Lo, Thresh.Hi, ScoreRegime)}
}
\value{
vector of scores
}
\description{
This function calculates metric scores based on a Thresholds data frame.
Can generate scores for 1/3/5 (ScoreRegime="135") or 0-100 (ScoreRegime="100").
}
\examples{
# Fish Metrics (generate values then score)
myIndex <- "MBSS.2005.Fish"
# Thresholds
thresh <- metrics_scoring
# get metric names for myIndex
(myMetrics.Fish <- as.character(droplevels(unique(thresh[thresh[,"Index.Name"]==myIndex,"Metric"]))))
# Taxa Data
myDF.Fish <- taxa_fish
Metrics.Fish <- metric.values(myDF.Fish, "SampleID", "fish", myMetrics.Fish, TRUE)
View(Metrics.Fish)
# SCORE
Metrics.Fish.Scores <- metric.scores(Metrics.Fish, myMetrics.Fish, "Index.Name", "FIBISTRATA", thresh)
# View Results
View(Metrics.Fish.Scores)

}
