% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{taxa_bugs_family}
\alias{taxa_bugs_family}
\title{taxa_bugs_family}
\format{
a dataframe with 2749 observations and 20 variables:
}
\source{
MBSS 2016 benthic macroinvertebrate saved as MSW family data.
}
\usage{
taxa_bugs_family
}
\description{
MSW benthic macroinvertebrate family taxa information from MBSS 2016 data.  Used as example for calculating metrics and index.
}
\keyword{datasets}
